+++ inputs/2020/14.txt
