+++ inputs/2020/3.txt
