+++ inputs/2020/6.txt
