+++ inputs/2020/7.txt
