+++ inputs/2020/8.txt
