+++ src/2020/1.js
@@ -1,6 +1,7 @@
-import { readLines } from '../../lib/read-input.js';
+import assert from 'assert';
+import { readLines } from '../lib/read-input.js';
 
-const numbers = readLines('2020', '01').map(Number);
+const numbers = readLines(2020, 1).map(Number);
 
 function part1() {
     for (const i of numbers) {
@@ -24,4 +25,11 @@ function part2() {
     }
 }
 
-export { part1, part2 };
+describe('2020 - Day 1', () => {
+    it('part1 is 545379', () => {
+        assert.strictEqual(part1(), 545379);
+    });
+    it('part2 is 257778836', () => {
+        assert.strictEqual(part2(), 257778836);
+    });
+});
