+++ src/2020/10.js
@@ -1,7 +1,8 @@
+import assert from 'assert';
 import _ from 'lodash';
-import { readLines } from '../../lib/read-input.js';
+import { readLines } from '../lib/read-input.js';
 
-const input = _(readLines('2020', '10')).map(Number).sortBy().value();
+const input = _(readLines(2020, 10)).map(Number).sortBy().value();
 input.push(input[input.length - 1] + 3);
 input.unshift(0);
 
@@ -38,4 +39,11 @@ function countPaths(start, cache) {
     return paths;
 }
 
-export { part1, part2 };
+describe('2020 - Day 10', () => {
+    it('part1 is 1836', () => {
+        assert.strictEqual(part1(), 1836);
+    });
+    it('part2 is 43406276662336', () => {
+        assert.strictEqual(part2(), 43406276662336);
+    });
+});
