+++ src/2020/13.js
@@ -1,6 +1,7 @@
-import { readLines } from '../../lib/read-input.js';
+import assert from 'assert';
+import { readLines } from '../lib/read-input.js';
 
-const [line1, line2] = readLines('2020', '13');
+const [line1, line2] = readLines(2020, 13);
 
 function part1() {
     const start = Number(line1);
@@ -33,4 +34,11 @@ function findEarliestTimestamp([[id, offset], ...rest], timeStamp = 0, dt = 1) {
     return rest.length ? findEarliestTimestamp(rest, t[0], t[1] - t[0]) : t[0];
 }
 
-export { part1, part2 };
+describe('2020 - Day 13', () => {
+    it('part1 is 207', () => {
+        assert.strictEqual(part1(), 207);
+    });
+    it('part2 is 530015546283687', () => {
+        assert.strictEqual(part2(), 530015546283687);
+    });
+});
