+++ src/2020/14.js
@@ -1,6 +1,7 @@
-import { readLines } from '../../lib/read-input.js';
+import assert from 'assert';
+import { readLines } from '../lib/read-input.js';
 
-const input = readLines('2020', '14').map((line) => line.split(' = '));
+const input = readLines(2020, 14).map((line) => line.split(' = '));
 const X = /X/g;
 const memRegex = /mem\[|\]/g;
 
@@ -19,4 +20,8 @@ function part1() {
     return mem.filter(Boolean).reduce((a, b) => a + b);
 }
 
-export { part1 };
+describe('2020 - Day 14', () => {
+    it('part1 is 6513443633260n', () => {
+        assert.strictEqual(part1(), 6513443633260n);
+    });
+});
