+++ src/2020/6.js
@@ -1,8 +1,9 @@
+import assert from 'assert';
 import _ from 'lodash';
-import { readLineGroups } from '../../lib/read-input.js';
+import { readLineGroups } from '../lib/read-input.js';
 
 function getInput() {
-    return _(readLineGroups('2020', '06'))
+    return _(readLineGroups(2020, 6))
         .map((lines) => {
             const allAnswers = lines.join(''); // [abc,def] -> abcdef
             const uniqueAnswers = new Set(allAnswers.split('')); // abcabcde -> Set(a,b,c,d,e)
@@ -23,4 +24,11 @@ function part2() {
     return getInput().sum();
 }
 
-export { part1, part2 };
+describe('2020 - Day 6', () => {
+    it('part1 is 6310', () => {
+        assert.strictEqual(part1(), 6310);
+    });
+    it('part2 is 3193', () => {
+        assert.strictEqual(part2(), 3193);
+    });
+});
