+++ src/2020/7.js
@@ -1,8 +1,9 @@
+import assert from 'assert';
 import _ from 'lodash';
-import { readLines } from '../../lib/read-input.js';
+import { readLines } from '../lib/read-input.js';
 
 const children = new Map(
-    readLines('2020', '07').map((line) => {
+    readLines(2020, 7).map((line) => {
         const [color, rest] = line.split(' bags contain ');
         const contents = rest
             .split(', ')
@@ -51,4 +52,11 @@ function collectBags([opened, ...rest], store) {
     return [opened].concat(bags.length ? collectBags(bags, store) : []);
 }
 
-export { part1, part2 };
+describe('2020 - Day 7', () => {
+    it('part1 is 213', () => {
+        assert.strictEqual(part1(), 213);
+    });
+    it('part2 is 38426', () => {
+        assert.strictEqual(part2(), 38426);
+    });
+});
