+++ src/2020/9.js
@@ -1,7 +1,8 @@
+import assert from 'assert';
 import _ from 'lodash';
-import { readLines } from '../../lib/read-input.js';
+import { readLines } from '../lib/read-input.js';
 
-const input = readLines('2020', '09').map(Number);
+const input = readLines(2020, 9).map(Number);
 
 function part1() {
     const entries = input.slice(25).entries();
@@ -35,4 +36,11 @@ function hasSum(arr, sumToFind) {
     return false;
 }
 
-export { part1, part2 };
+describe('2020 - Day 9', () => {
+    it('part1 is 23278925', () => {
+        assert.strictEqual(part1(), 23278925);
+    });
+    it('part2 is 4011064', () => {
+        assert.strictEqual(part2(), 4011064);
+    });
+});
